# [bf2/D]: Improve performance and code elegance