Results:

|        | LDC bench.b           | LDC mandel.b       | GDC bench.b       | GDC mandel.b           | DMD bench.b           | DMD mandel.b       |
|--------|-----------------------|--------------------|-------------------|------------------------|-----------------------|--------------------|
| Before | 2.79s / 1.0Mb         | 36.06s / 2.8Mb     | **2.19s** / 2.8Mb | 31.16s / 3.7Mb         | 5.32s / 1.5Mb         | 79.70s / 3.4Mb     |
| After  | **1.87s** / **0.9Mb** | **22.57s** / 2.8Mb | 2.54s / **2.7Mb** | **27.46s** / **3.5Mb** | **4.92s** / **1.4Mb** | **60.18s** / 3.4Mb |

For reference, here are the C++ results for my machine:

|     | g++ bench.b | g++ mandel.b |
|-----|-------------|--------------|
| C++ | 2.19s / 1.5Mb      | 20.50s / 3.3Mb       |

```
$ g++ -O3 -o bin_cpp bf.cpp && ../xtime.rb ./bin_cpp ./bench.b 
2.19s, 1.5Mb
$ g++ -O3 -o bin_cpp bf.cpp && ../xtime.rb ./bin_cpp ./mandel.b
20.50s, 3.3Mb
```

D results before:
```
$ ldc2 -of bin_d_ldc -O5 -release bf.d && ../xtime.rb ./bin_d_ldc ./bench.b
2.79s, 1.0Mb
$ ldc2 -of bin_d_ldc -O5 -release bf.d && ../xtime.rb ./bin_d_ldc ./mandel.b
36.06s, 2.8Mb

$ gdc-5 -o bin_d_gdc -O3 -frelease -finline bf.d && ../xtime.rb ./bin_d_gdc ./bench.b
2.19s, 2.8Mb
$ gdc-5 -o bin_d_gdc -O3 -frelease -finline bf.d && ../xtime.rb ./bin_d_gdc ./mandel.b
31.16s, 3.7Mb

$ dmd -ofbin_d -O -inline -release bf.d && ../xtime.rb ./bin_d ./bench.b
5.32s, 1.5Mb
$ dmd -ofbin_d -O -inline -release bf.d && ../xtime.rb ./bin_d ./mandel.b
79.70s, 3.4Mb
```

D results after:
```
$ ldc2 -of bin_d_ldc -O5 -release bf.d && ../xtime.rb ./bin_d_ldc ./bench.b
1.87s, 0.9Mb
$ ldc2 -of bin_d_ldc -O5 -release bf.d && ../xtime.rb ./bin_d_ldc ./mandel.b
22.57s, 2.8Mb

$ gdc-5 -o bin_d_gdc -O3 -frelease -finline bf.d && ../xtime.rb ./bin_d_gdc ./bench.b
2.54s, 2.7Mb
$ gdc-5 -o bin_d_gdc -O3 -frelease -finline bf.d && ../xtime.rb ./bin_d_gdc ./mandel.b
27.46s, 3.5Mb

$ dmd -ofbin_d -O -inline -release bf.d && ../xtime.rb ./bin_d ./bench.b
4.92s, 1.4Mb
$ dmd -ofbin_d -O -inline -release bf.d && ../xtime.rb ./bin_d ./mandel.b
60.18s, 3.4Mb
```
(The actual program output is omitted for brevity.)

As you can see, the only regression is GDC bench.b, but apart from that it's a clear improvement across the board.

Some info about the test environment:
```
$ ldc2 --version
LDC - the LLVM D compiler (1.0.0):
  based on DMD v2.070.2 and LLVM 3.8.0
  built with LDC - the LLVM D compiler (1.0.0)
  Default target: x86_64-unknown-linux-gnu
  Host CPU: ivybridge
  http://dlang.org - http://wiki.dlang.org/LDC

  Registered Targets:
    x86    - 32-bit X86: Pentium-Pro and above
    x86-64 - 64-bit X86: EM64T and AMD64

$ gdc-5 --version
gdc-5 (Ubuntu 5.3.1-14ubuntu2.1) 5.3.1 20160413
Copyright (C) 2015 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

$ g++ --version
g++ (Ubuntu 5.3.1-14ubuntu2.1) 5.3.1 20160413
Copyright (C) 2015 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

$ dmd --version
DMD64 D Compiler v2.070.0
Copyright (c) 1999-2015 by Digital Mars written by Walter Bright

$ uname -a
Linux zombinedev-desktop 4.5.0-040500-generic #201603140130 SMP Mon Mar 14 05:32:22 UTC 2016 x86_64 x86_64 x86_64 GNU/Linux

$ cat /proc/cpuinfo | grep -e "model name" | head -1
model name	: Intel(R) Core(TM) i5-3450 CPU @ 3.10GHz
```