@kostya 2016-06-25
thanks, i ll try it soon, no really understand why it faster now, seems nothing valuable changed