@ZombineDev 2016-06-25
Well I tried to keep the algorithm as it is (even though it was hard to resist :D). Most of the changes are:
+ removed everything unnecessary like `StringIterator`
+ made program `Op`s `immutable` (instead of mutable)
+ used static `foreach` to make the dispatch faster in `run`
+ changed classes to structs because polymorphism wasn't needed (everything was `final` before).
+ made `run` and `parse` with arguments static. I would prefer them to be free functions and to remove `Program` all together, but I didn't want to deviate much from the original.