@ZombineDev 2016-06-25
Some future optimizations would be:
+ Pre-allocate the tape (e.g. change `int[] tape = [0]` to `int[] tape = new int[1024]`) and reduce moveNext only to `pos++`, but I wasn't sure if this would be a fair. Also it didn't make a big difference: 22.57s vs 21.11s (for LDC, mandel.b)
+ Use `core.stdc.stdio.putchar()` instead of `std.stdio.write()` + `std.stdio.File.flush()` (i.e. disable flushing), but again I wasn't sure if it's fair to disable flushing, and secondly, it didn't bring consistently better results - for some strange reason it was faster only combination with the above optimization (22.57s-> 21.11s -> 20.07s), but it was slower on its own (22.57s->23.47s).