@ZombineDev 2016-06-25
I added C++ times for reference to table above.