@jadbox 2016-06-26
How did your changes end up reducing the memory footprint?