@ZombineDev 2016-06-27
I wouldn't (and don't) advertise that it reduces memory usage, because during testing it was often the case that the results would vary between +/- 0.1MB, though it was relatively stable. As you can see, it depends mostly on the compiler - e.g. newer versions of druntime may have different array growth factors, also memory allocations could be elided by the compiler (I suspect that LDC might be doing that).
My biggest motivation for this change is that the run time is noticeably reduced for the majority of cases (compilers and test cases) and that code is shorter / easier to follow.