@ZombineDev 2016-06-27
@jadbox Oh, I forgot to mention, since I changed all classes to structs, those objects are now placed on the stack (instead of the heap), so this surely accounts for some reduction of memory usage, albeit a tiny one. (I'm talking about `Program` and `Tape`.)